api/requirements.txt
fastapi==0.100.0
sqlmodel==0.0.8
psycopg2-binary==2.9.7
uvicorn==0.22.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
cryptography==41.0.3
passlib[bcrypt]==1.7.4
alembic==1.11.1
pytest==7.4.1
httpx==0.24.1
